+++ src/cpu/x64/rnn/jit_uni_rnn_cell_postgemm_bwd.hpp
@@ -45,8 +45,9 @@ protected:
     using Vmm = typename cpu_isa_traits_t<isa>::Vmm;
     static constexpr int vlen = cpu_isa_traits_t<isa>::vlen;
     static constexpr int hstate_dt_size = sizeof(float);
-    const int vlen_scratch
-            = vlen / (sizeof(float) / types::data_type_size(scratch_data_t));
+    const int vlen_scratch = vlen
+            / static_cast<int>(
+                    sizeof(float) / types::data_type_size(scratch_data_t));
     const int gate_dt_size
             = static_cast<int>(types::data_type_size(scratch_data_t));
     const int scratch_dt_size
